requirements.txt
torch~=2.1.1
detecto~=1.2.2
matplotlib~=3.8.2
torchvision~=0.16.1
gdown~=4.7.1
pillow~=10.1.0